# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: API Integration
# during dev, should point to your local machine
# host: localhost:5040
host: 52.39.212.226:5054
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/addNewDrug:
    # binds a127 app logic to a route
    x-swagger-router-controller: inventory_ctrl
    post:
      tags: ['Inventory Module']
      description: Add New Drug
      summary:  Add New Drug
      # used as the method name of the controller
      operationId: addNewDrug
      parameters:
        - name: body
          in: body
          description:  Add New Drug
          schema:
           $ref: '#/definitions/addNewDrug'
          required: true       
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string               
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/getDrugList:
    # binds a127 app logic to a route
    x-swagger-router-controller: inventory_ctrl
    get:
      tags: ['Inventory Module']
      description: Get All Drug
      summary:  Get All Drug
      # used as the method name of the controller
      operationId: getDrugList      
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                    
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/getDrugs:
    # binds a127 app logic to a route
    x-swagger-router-controller: inventory_ctrl
    get:
      tags: ['Inventory Module']
      description: Get All Drug
      summary:  Get All Drug
      # used as the method name of the controller
      operationId: getDrugs      
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                       
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/deleteDrug/{id}/{to}:
    # binds a127 app logic to a route
    x-swagger-router-controller: inventory_ctrl
    delete:
      tags: ['Inventory Module']
      description: Delete Drug By Id
      summary: Delete Drug By Id
      # used as the method name of the controller
      operationId: deleteDrug
      parameters:
        - name: id
          in: path
          description: drug id
          required: true
          type: string
        - name: to
          in: path
          description: drugDetail id
          required: true
          type: string
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/download:
    # binds a127 app logic to a route
    x-swagger-router-controller: patient_ctrl
    post:
      tags: ['Patient Module']
      description: Add Follow Up Details
      summary: Add Follow Up Details
      # used as the method name of the controller
      operationId: download
      produces:
          - application/octet-stream
      parameters:
        - name: body
          in: body
          description: Add Follow Up Details
          schema:
           $ref: '#/definitions/download'
          required: true  
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                     
      responses:
        # "200":
        #   description: Success
        #   schema:
        #     # a pointer to a definition
        #     $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"         
...
definitions:                         
...
  download:
    type: object
    properties:
      fileName:
        type: string   
...
  appResponse:
    # required:
    #   - code
    properties:
      code:
        type: number
      status:
        type: string        
      message:
        type: string        
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  # by sunny on 15-april-2017
  addNewDrug:
    type: object
    properties:
      drugName:
        type: string
      _id:
        type: string
      drugDetails:
       type: object
       items:
        type: object
        properties:
          quantity:
           type: number
          dosage:
           type: string 
          updateQuantity:
           type: string 
...